Added Attractions and Events page.
Added both pages with random names for data. Of course added an event for the one and only Ian~ Going to try and add a Google Maps location (using Six Flags) in the About Me page.

I also need to fix why the Scroll isn't working in the Home Page so I can finish it up.